Refactor out puppeteer utils that don't depend on the test closure
